Add "extra" parameter to handle external dependencies.
The extra parameter will take a string or an array of strings
that each refer to a filename. These files are configuration
files related to the stream we're processing; any change in the
configuration files will cause *all* of the files to be rebuilt.

You can, for instance, add gulpfile.js to extra, causing all files
to be rebuilt when gulpfile.js is touched. Alternately, if your
lint configuration file is changed, gulp-newer can re-lint all
source files.